Extract request logic to its own class
Keep client as clean as possible.
