+++ lib/google_maps/directions/client.rb
@@ -1,8 +1,5 @@
 # frozen_string_literal: true
 
-require 'addressable'
-require 'faraday'
-
 module GoogleMaps
   module Directions
     # Target functionality
@@ -19,38 +16,17 @@ module GoogleMaps
     class Client
       extend Forwardable
 
-      attr_accessor :api_key, :adaptor, :default_options
-
-      def_delegator 'GoogleMaps::Directions', :config
-
-      def initialize(adaptor: Faraday)
-        self.adaptor = adaptor
-        self.api_key = config.api_key
-        self.default_options = config.default_options
-      end
+      def_delegator 'GoogleMaps::Directions::Request', :get
 
       def directions(origin:, destination:, **options)
-        response = request(origin, destination, **options)
-        result = GoogleMaps::Directions::Result.new(response.body)
+        response = get(origin, destination, **options)
+        result = Result.new(response.body)
         return result if result.success?
 
         handle_error(result)
       end
 
       private
-
-      def request(origin, destination, **options)
-        uri.query_values = default_options.merge({ key: api_key, origin: origin, destination: destination, **options })
-        adaptor.get(uri)
-      end
-
-      def uri
-        @uri ||= Addressable::URI.parse(directions_path)
-      end
-
-      def directions_path
-        GoogleMaps::Directions.path
-      end
 
       def handle_error(result)
         case result['status']
