+++ lib/google_maps/directions/errors.rb
@@ -4,8 +4,12 @@
 #
 module GoogleMaps
   module Directions
+    # Base error for lib
+    #
     class Error < StandardError; end
 
+    # Raised when status REQUEST_DENIED in API response
+    #
     class RequestDeniedError < Error; end
   end
 end
